pyproject.toml
[tool.poetry]
name = "hexlet-friends"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
Django = "^4.2.14"
python-dotenv = "^0.20.0"
requests = "^2.31.0"
gunicorn = "^20.1.0"
dj-database-url = "^0.5"
whitenoise = "^6.2.0"
python-dateutil = "^2.8.2"
APScheduler = "^3.9.1"
django-crispy-forms = "^1.14.0"
django-cte = "^1.2.1"
psycopg2 = { version = "^2.8", optional = true }
psycopg2-binary = { version = "^2.9.3", optional = true }
PyJWT = "^2.4.0"
cryptography = "^41.0.0"
crispy-bootstrap5 = "^0.6"
django-extensions = "^3.2.1"
black = "^23.3.0"
django-debug-toolbar = "^3.8.1"
django-mathfilters = "^1.0.0"
django-filter = "^23.2"
sentry-sdk = "^1.29.2"

[tool.poetry.dev-dependencies]
Faker = "^13.15.1"
coverage = "^6.4.2"
pre-commit = "^3.5.0"
pyparsing = "^3.1.1"
pydot = "^1.4.2"


[tool.poetry.extras]
psycopg2 = ["psycopg2"]
psycopg2-binary = ["psycopg2-binary"]


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry>=1.0.5"]
build-backend = "poetry.masonry.api"
